+++ src/cores/api/api-service.ts
@@ -11,9 +11,9 @@
  * @copyright   (C) lemoncloud.io 2019 - All Rights Reserved.
  */
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
-import { $engine, _log, _inf, _err, $U } from '../engine/';
-import { loadJsonSync } from '../tools/shared';
-import { GETERR } from '../common/test-helper';
+import { $engine, _log, _inf, _err, $U } from '../../engine/';
+import { loadJsonSync } from '../../tools/shared';
+import { GETERR } from '../../common/test-helper';
 import fs from 'fs';
 const NS = $U.NS('APIS', 'green'); // NAMESPACE TO BE PRINTED.
