+++ src/cores/api/index.ts
@@ -1,0 +1,10 @@
+/**
+ * `cores/api/index.tx`
+ * - default index to export all for api-service
+ *
+ * @author      Steve Jung <steve@lemoncloud.io>
+ * @date        2022-05-13 initial file to group
+ *
+ * @copyright (C) lemoncloud.io 2022 - All Rights Reserved.
+ */
+export * from './api-service';
